Move `lua` package into the `api` package.
This should make it easier in the future if the API were to ever be moved into it's own subproject and it makes sense since the entire "Lua envoirment" is just an "api" now.
