+++ src/main/kotlin/xyz/galaxyy/lualink/api/lua/LuaAddons.kt
@@ -1,4 +1,4 @@
-package xyz.galaxyy.lualink.lua
+package xyz.galaxyy.lualink.api.lua
 
 import org.luaj.vm2.LuaTable
 import org.luaj.vm2.lib.OneArgFunction
