+++ src/main/kotlin/xyz/galaxyy/lualink/api/lua/LuaImport.kt
@@ -1,4 +1,4 @@
-package xyz.galaxyy.lualink.lua
+package xyz.galaxyy.lualink.api.lua
 
 import org.luaj.vm2.LuaValue
 import org.luaj.vm2.lib.OneArgFunction
