+++ src/main/kotlin/xyz/galaxyy/lualink/commands/LoadedScriptParser.kt
@@ -5,7 +5,7 @@ import cloud.commandframework.arguments.parser.ArgumentParser
 import cloud.commandframework.context.CommandContext
 import cloud.commandframework.exceptions.parsing.NoInputProvidedException
 import xyz.galaxyy.lualink.api.LuaScript
-import xyz.galaxyy.lualink.lua.LuaScriptManager
+import xyz.galaxyy.lualink.api.lua.LuaScriptManager
 import java.util.*
 
 internal class LoadedScriptParser<C: Any>(private val scriptManager: LuaScriptManager)  : ArgumentParser<C, LuaScript> {
