+++ src/main/kotlin/xyz/galaxyy/lualink/commands/LuaLinkCommands.kt
@@ -12,7 +12,7 @@ import org.bukkit.entity.Player
 import org.luaj.vm2.lib.jse.JsePlatform
 import xyz.galaxyy.lualink.LuaLink
 import xyz.galaxyy.lualink.api.LuaScript
-import xyz.galaxyy.lualink.lua.LuaScriptManager
+import xyz.galaxyy.lualink.api.lua.LuaScriptManager
 import java.io.File
 
 @Suppress("unused")
