+++ src/main/kotlin/xyz/galaxyy/lualink/listeners/ServerLoadListener.kt
@@ -3,7 +3,7 @@ package xyz.galaxyy.lualink.listeners
 import org.bukkit.event.EventHandler
 import org.bukkit.event.Listener
 import org.bukkit.event.server.ServerLoadEvent
-import xyz.galaxyy.lualink.lua.LuaScriptManager
+import xyz.galaxyy.lualink.api.lua.LuaScriptManager
 
 internal class ServerLoadListener(private val scriptManager: LuaScriptManager) : Listener {
     @EventHandler
